Split out submission handler and debug protocol
