+++ flowrlib/src/debugger_protocol.rs
@@ -1,54 +1,17 @@
-#[cfg(feature = "debugger")]
 use serde_json::Value;
 
 use flowcore::errors::*;
-#[cfg(feature = "debugger")]
 use flowcore::model::input::Input;
-#[cfg(feature = "metrics")]
-use flowcore::model::metrics::Metrics;
-#[cfg(feature = "debugger")]
 use flowcore::model::output_connection::OutputConnection;
-#[cfg(feature = "debugger")]
 use flowcore::model::runtime_function::RuntimeFunction;
-use flowcore::model::submission::Submission;
 
-#[cfg(feature = "debugger")]
 use crate::block::Block;
-#[cfg(feature = "debugger")]
 use crate::debug_command::DebugCommand;
-#[cfg(feature = "debugger")]
 use crate::job::Job;
 use crate::run_state::RunState;
-#[cfg(feature = "debugger")]
 use crate::run_state::State;
 
-/// Programs linking `flowrlib` that wish to submit a flow for execution via a `Submission` and
-/// then track it's execution (such as a CLI or a UI) should implement this trait.
-pub trait SubmissionProtocol {
-    /// Execution of the flow is starting
-    fn flow_execution_starting(&mut self) -> Result<()>;
-
-    /// The `Coordinator` executing the flow periodically will check if there has been a request
-    /// to enter the debugger.
-    #[cfg(feature = "debugger")]
-    fn should_enter_debugger(&mut self) -> Result<bool>;
-
-    /// The execution of the flow has ended
-    #[cfg(feature = "metrics")]
-    fn flow_execution_ended(&mut self, state: &RunState, metrics: Metrics) -> Result<()>;
-    /// The flow has ended
-    #[cfg(not(feature = "metrics"))]
-    fn flow_execution_ended(&mut self, state: &RunState) -> Result<()>;
-
-    /// Wait for a `Submission` to be sent to the `Coordinator` for execution
-    fn wait_for_submission(&mut self) -> Result<Option<Submission>>;
-
-    /// The thread or process running the `Coordinator` to execute the flow is about to exit
-    fn coordinator_is_exiting(&mut self, result: Result<()>) -> Result<()>;
-}
-
 /// Programs that wish to offer a debugger (such as a CLI or UI) should implement this protocol
-#[cfg(feature = "debugger")]
 pub trait DebuggerProtocol {
     /// Start the debugger - which swallows the first message to initialize the connection
     fn start(&mut self);
